+++ packages/gasket-docs-plugin/test/index.test.js
@@ -1,0 +1,9 @@
+const plugin = require('../');
+const assume = require('assume');
+
+describe('Plugin', function () {
+
+  it('is named correctly', function () {
+    assume(plugin.name).equals('docs');
+  });
+});